pyproject.toml
[project]
name = "OMNI-OPERATOR-V1"
version = "0.1.0"
description = "Autonomiczna fabryka dystrybucji treści - Gemini Hackathon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai==1.39.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "qdrant-client>=1.16.2",
    "langfuse<3.0.0",
    "moviepy>=2.2.1",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.4.0",
    "torch",
    "torchvision",
    "torchaudio",
    "google-genai>=1.56.0",
]

[tool.uv]
managed = true
package = false
